pyproject.toml
[tool.poetry]
name = "repl_python3_discord-song-bot"
version = "0.1.0"
description = ""
authors = ["pedersendane <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.20.0"
youtube_dl = "^2021.12.17"
replit = "^3.2.4"
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
